New controller for booking visual
