+++ backend-api/GetawaysNow/src/main/java/com/example/GetawaysNow/booking/BookingController.java
@@ -1,69 +1,34 @@
 package com.example.GetawaysNow.booking;
 
-import java.util.List;
-
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
-import org.springframework.web.bind.annotation.PutMapping;
-import org.springframework.web.bind.annotation.RequestBody;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RestController;
-import org.springframework.stereotype.Controller;
-
-
-import jakarta.validation.Valid;
+import org.springframework.web.bind.annotation.RequestParam;
 
 @Controller
-public class BookingController {
-
+public class BookingController{
     private final BookingService bookingService;
 
     @Autowired
-    public BookingController(BookingService bookingService) {
+    public BookingController(BookingService bookingService){
         this.bookingService = bookingService;
     }
 
-    @PostMapping("/profile/{profileId}/listing/{listingId}")
-    public ResponseEntity<Booking> createBooking(
-            @PathVariable Long profileId,
-            @PathVariable Long listingId,
-            @Valid @RequestBody Booking bookingDetails) {
-
-        Booking newBooking = bookingService.createBooking(profileId, listingId, bookingDetails);
-        return ResponseEntity.ok(newBooking);
+    @GetMapping("/checkout/profile/{profileId}/listing/{listingId}")
+    public String showCheckoutForm(@PathVariable Long profileId, @PathVariable Long listingId, Model model){
+        model.addAttribute("booking", new Booking());
+        model.addAttribute("profileId", profileId);
+        model.addAttribute("listingId", listingId);
+        return "checkout";
     }
 
-    @GetMapping
-    public ResponseEntity<List<Booking>> getAllBookings() {
-        List<Booking> bookings = bookingService.getAllBookings();
-        return ResponseEntity.ok(bookings);
+    @PostMapping("/booking/create")
+    public String createBooking(Booking booking, @RequestParam Long profileId, @RequestParam Long listingId){
+        bookingService.createBooking(profileId, listingId, booking);
+        return "redirect:/profile/" + profileId;
     }
 
-    @GetMapping("/{id}")
-    public ResponseEntity<Booking> getBookingById(@PathVariable Long id) {
-        Booking booking = bookingService.getBookingById(id);
-        return ResponseEntity.ok(booking);
-    }
-
-    @GetMapping("/profile/{profileId}")
-    public ResponseEntity<List<Booking>> getBookingsByProfileId(@PathVariable Long profileId) {
-        List<Booking> bookings = bookingService.getBookingsByProfileId(profileId);
-        return ResponseEntity.ok(bookings);
-    }
-
-    @PutMapping("/{id}")
-    public ResponseEntity<Booking> updateBooking(@PathVariable Long id, @Valid @RequestBody Booking bookingDetails) {
-        Booking updatedBooking = bookingService.updateBooking(id, bookingDetails);
-        return ResponseEntity.ok(updatedBooking);
-    }
-
-    @DeleteMapping("/{id}")
-    public ResponseEntity<Void> deleteBooking(@PathVariable Long id) {
-        bookingService.deleteBooking(id);
-        return ResponseEntity.noContent().build();
-    }
 }
