+++ backend-api/GetawaysNow/src/main/java/com/example/GetawaysNow/booking/BookingRestController.java
@@ -1,0 +1,69 @@
+package com.example.GetawaysNow.booking;
+
+import java.util.List;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.PutMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+
+import jakarta.validation.Valid;
+
+@RestController
+@RequestMapping("/api/bookings")
+public class BookingRestController {
+
+    private final BookingService bookingService;
+
+    @Autowired
+    public BookingRestController(BookingService bookingService) {
+        this.bookingService = bookingService;
+    }
+
+    @PostMapping("/profile/{profileId}/listing/{listingId}")
+    public ResponseEntity<Booking> createBooking(
+            @PathVariable Long profileId,
+            @PathVariable Long listingId,
+            @Valid @RequestBody Booking bookingDetails) {
+
+        Booking newBooking = bookingService.createBooking(profileId, listingId, bookingDetails);
+        return ResponseEntity.ok(newBooking);
+    }
+
+    @GetMapping
+    public ResponseEntity<List<Booking>> getAllBookings() {
+        List<Booking> bookings = bookingService.getAllBookings();
+        return ResponseEntity.ok(bookings);
+    }
+
+    @GetMapping("/{id}")
+    public ResponseEntity<Booking> getBookingById(@PathVariable Long id) {
+        Booking booking = bookingService.getBookingById(id);
+        return ResponseEntity.ok(booking);
+    }
+
+    @GetMapping("/profile/{profileId}")
+    public ResponseEntity<List<Booking>> getBookingsByProfileId(@PathVariable Long profileId) {
+        List<Booking> bookings = bookingService.getBookingsByProfileId(profileId);
+        return ResponseEntity.ok(bookings);
+    }
+
+    @PutMapping("/{id}")
+    public ResponseEntity<Booking> updateBooking(@PathVariable Long id, @Valid @RequestBody Booking bookingDetails) {
+        Booking updatedBooking = bookingService.updateBooking(id, bookingDetails);
+        return ResponseEntity.ok(updatedBooking);
+    }
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> deleteBooking(@PathVariable Long id) {
+        bookingService.deleteBooking(id);
+        return ResponseEntity.noContent().build();
+    }
+}
